pyproject.toml
[project]
name = "qwick-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "python-telegram-bot>=21.11.1",
]
